Correção dos nomes dos pacotes de pagamento
