+++ pagamento/src/main/java/br/com/academy/pagamento/PagamentoApplication.java
@@ -1,0 +1,13 @@
+package br.com.academy.pagamento;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PagamentoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PagamentoApplication.class, args);
+	}
+
+}
