+++ pagamento/src/main/java/br/com/academy/pagamento/entity/ProdutoVenda.java
@@ -1,0 +1,39 @@
+package br.com.academy.pagamento.entity;
+
+import java.io.Serializable;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "produto_venda")
+@Data
+@NoArgsConstructor
+public class ProdutoVenda implements Serializable {
+
+	private static final long serialVersionUID = 7782449934590586015L;
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	
+	@Column(name = "id_produto")
+	private Long idProduto;
+	
+	@Column(name = "quantidade", length = 10)
+	private Integer quantidade;
+	
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "id_venda")
+	private Venda venda;
+}
