+++ pagamento/src/main/java/br/com/academy/pagamento/exception/NotFoundException.java
@@ -1,0 +1,26 @@
+package br.com.academy.pagamento.exception;
+
+public class NotFoundException extends RuntimeException {
+
+	private static final long serialVersionUID = -182478625499114603L;
+
+	public NotFoundException() {
+		super();
+	}
+
+	public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+		super(message, cause, enableSuppression, writableStackTrace);
+	}
+
+	public NotFoundException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public NotFoundException(String message) {
+		super(message);
+	}
+
+	public NotFoundException(Throwable cause) {
+		super(cause);
+	}
+}
