+++ pagamento/src/main/java/br/com/academy/pagamento/repository/ProdutoRepository.java
@@ -1,0 +1,11 @@
+package br.com.academy.pagamento.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.academy.pagamento.entity.Produto;
+
+@Repository
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+
+}
