+++ pagamento/src/main/java/br/com/academy/pagamento/repository/VendaRepository.java
@@ -1,0 +1,11 @@
+package br.com.academy.pagamento.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.academy.pagamento.entity.Venda;
+
+@Repository
+public interface VendaRepository extends JpaRepository<Venda, Long> {
+
+}
