+++ pagamento/src/main/java/br/com/academy/pagamento/service/ProdutoService.java
@@ -1,0 +1,5 @@
+package br.com.academy.pagamento.service;
+
+public class ProdutoService {
+
+}
