--- pagamento/src/main/java/br/com/academy/pagemento/PagementoApplication.java
@@ -1,13 +1,0 @@
-package br.com.academy.pagemento;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PagementoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PagementoApplication.class, args);
-	}
-
-}
