--- pagamento/src/main/java/br/com/academy/pagemento/entity/Produto.java
@@ -1,26 +1,0 @@
-package br.com.academy.pagemento.entity;
-
-import java.io.Serializable;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "produto")
-@Data
-@NoArgsConstructor
-public class Produto implements Serializable {
-	
-	private static final long serialVersionUID = -1512332340110143752L;
-	
-	@Id
-	private Long id;
-	
-	@Column(name = "estoque", nullable = false, length = 10)
-	private Integer estoque;
-}
