--- pagamento/src/main/java/br/com/academy/pagemento/model/ProdutoVO.java
@@ -1,25 +1,0 @@
-package br.com.academy.pagemento.model;
-
-import java.io.Serializable;
-
-import org.springframework.hateoas.RepresentationModel;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@EqualsAndHashCode(callSuper = false)
-public class ProdutoVO extends RepresentationModel<ProdutoVO> implements Serializable {
-
-	private static final long serialVersionUID = -3426203315108220331L;
-
-	private Long id;
-	
-	private Integer estoque;
-}
