--- pagamento/src/main/java/br/com/academy/pagemento/repository/ProdutoRepository.java
@@ -1,11 +1,0 @@
-package br.com.academy.pagemento.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import br.com.academy.pagemento.entity.Produto;
-
-@Repository
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
-
-}
