--- pagamento/src/main/java/br/com/academy/pagemento/service/ProdutoService.java
@@ -1,5 +1,0 @@
-package br.com.academy.pagemento.service;
-
-public class ProdutoService {
-
-}
