--- pagamento/src/main/java/br/com/academy/pagemento/service/exception/NotFoundException.java
@@ -1,29 +1,0 @@
-package br.com.academy.pagemento.service.exception;
-
-
-public class NotFoundException extends RuntimeException {
-
-	private static final long serialVersionUID = -2240718802847484551L;
-
-	public NotFoundException() {
-		super();
-	}
-
-	public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-		super(message, cause, enableSuppression, writableStackTrace);
-	}
-
-	public NotFoundException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public NotFoundException(String message) {
-		super(message);
-	}
-
-	public NotFoundException(Throwable cause) {
-		super(cause);
-	}
-	
-	
-}
